src/backend/requirements.txt
# defaul
redis
types-redis
uvicorn 
pytest
httpx

# env
pydantic-settings

# uvicorn
multiprocess

# Fastapi
fastapi
python-socketio

# database
sqlmodel
sqlalchemy
mysql-connector-python

alembic

# logger
loguru